+++ backend/controller/user_controller.py
@@ -13,7 +13,7 @@ user_blueprint = Blueprint('user', __name__)
 
 class UserController(MethodView):
     """
-    Controller for managing user-related actions like creating, updating, and deleting users,
+    controller for managing user-related actions like creating, updating, and deleting users,
     managing login sessions, and handling user files.
     """
 
@@ -90,6 +90,8 @@ class UserController(MethodView):
 
         :return: JSON response containing the status message and status code.
         """
+        if not request.is_json:
+            return jsonify({'error': 'Request data must be in JSON format'}), 400
         user_data = request.get_json()
         result = self.user_service.create_user(user_data)
         return jsonify(result.message), result.status_code
@@ -102,6 +104,8 @@ class UserController(MethodView):
 
         :return: JSON response containing the status message and status code.
         """
+        if not request.is_json:
+            return jsonify({'error': 'Request data must be in JSON format'}), 400
         user_data = request.get_json()
         result = self.user_service.update_user(user_data)
         return jsonify(result.message), result.status_code
@@ -115,6 +119,8 @@ class UserController(MethodView):
 
         :return: JSON response containing the status message and status code.
         """
+        if not request.is_json:
+            return jsonify({'error': 'Request data must be in JSON format'}), 400
         username_data = request.get_json()
         result = self.user_service.delete_user(username_data['username'])
         return jsonify(result.message), result.status_code
@@ -125,6 +131,8 @@ class UserController(MethodView):
 
         :return: JSON response containing the status message and status code.
         """
+        if not request.is_json:
+            return jsonify({'error': 'Request data must be in JSON format'}), 400
         credentials = request.get_json()
         result = self.auth_service.login(credentials['username'], credentials['password'])
         if not result.is_successful:
@@ -142,16 +150,18 @@ class UserController(MethodView):
         return jsonify(result.message), result.status_code
 
     @jwt_required()
-    @check_access(roles=[UserRole.ADMIN])
+
     def reset_password(self):
         """
         Resets the password for a user based on the provided JSON data.
 
         :return: JSON response with the result of the password reset attempt.
         """
+        if not request.is_json:
+            return jsonify({'error': 'Request data must be in JSON format'}), 400
         credentials = request.get_json()
         result = self.auth_service.reset_password(credentials['username'], credentials['password'])
-        return jsonify(result), result.status_code
+        return jsonify(result.message), result.status_code
 
     @jwt_required()
     def get_profile(self):
